accept release = 0 대입 후 SendPacket에 의한 release 오작동 해결 SendPost 후 결과처리 도달까지 시간 모니터링

+++ User.h
@@ -36,7 +36,7 @@ public:
 	unsigned short sector_x;
 	unsigned short sector_y;
 
-	unsigned int last_recv_time;
+	unsigned long long last_recv_time;
 
 
 };
